Move data grid helper classes to a separate file so forms editor would appear (I am going to commit violence against Microsoft).
