+++ GoodFilmsApp/DataGridWindow.cs
@@ -8,36 +8,6 @@ using ControllerLibrary;
 
 namespace GoodFilmsApp
 {
-    public class CStudioData
-    {
-        public int Id { get; set; }
-        public string Studio { get; set; }
-        public bool Chosen { get; set; }
-    }
-    public class CGenreData
-    {
-        public int Id { get; set; }
-        public string Genre { get; set; }
-        public bool Chosen { get; set; }
-    }
-    public class CDirectorData
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public bool Chosen { get; set; }
-    }
-    public class CAgeRatingData
-    {
-        public int Id { get; set; }
-        public string Rating { get; set; }
-        public bool Chosen { get; set; }
-    }
-    public class CLanguageData
-    {
-        public int Id { get; set; }
-        public string Language { get; set; }
-        public bool Chosen { get; set; }
-    }
     public partial class DataGridWindow : Form
     {
         //// Lists to store different models
